backend/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Giovane Costa <gigio.coder@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
celery = "^5.3.6"
sqlalchemy = "^2.0.27"
psycopg2 = "^2.9.9"
python-decouple = "^3.8"
redis = "^5.0.1"
python-multipart = "^0.0.9"
alembic = "^1.13.1"
fastapi-mail = "^1.4.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
ignore_missing_imports = true
